fix(tokenize): move tests to main file
